kindle_screen_shot_for_linux/requirements.txt
pillow
mss
pynput
PyQt5